+++ boilerplate/backend/python-drf/app/config/emailpassword.py
