+++ boilerplate/backend/python-drf/app/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for example project.
+Django settings for app project.
 
 Generated by 'django-admin startproject' using Django 3.2.14.
 
@@ -10,6 +10,11 @@ For the full list of settings and their values, see
 https://docs.djangoproject.com/en/3.2/ref/settings/
 """
 
+from corsheaders.defaults import default_headers
+from typing import List
+from supertokens_python import get_all_cors_headers
+from supertokens_python import init
+from app.config import supertokens_config, app_info, framework, recipe_list
 from pathlib import Path
 
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
@@ -54,7 +59,7 @@ MIDDLEWARE = [
     'supertokens_python.framework.django.django_middleware.middleware'
 ]
 
-ROOT_URLCONF = 'example.urls'
+ROOT_URLCONF = 'app.urls'
 
 TEMPLATES = [
     {
@@ -72,7 +77,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'example.wsgi.application'
+WSGI_APPLICATION = 'app.wsgi.application'
 
 
 # Database
@@ -130,10 +135,6 @@ STATIC_URL = '/static/'
 DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
 
 
-from example.config import supertokens_config, app_info, framework, recipe_list
-
-from supertokens_python import init
-
 init(
     supertokens_config=supertokens_config,
     app_info=app_info,
@@ -142,10 +143,6 @@ init(
     mode='wsgi'
 )
 
-
-from supertokens_python import get_all_cors_headers
-from typing import List
-from corsheaders.defaults import default_headers
 
 CORS_ORIGIN_WHITELIST = [
     "http://localhost:3000"
@@ -160,4 +157,3 @@ CORS_ALLOWED_ORIGINS = [
 CORS_ALLOW_HEADERS: List[str] = list(default_headers) + [
     "Content-Type"
 ] + get_all_cors_headers()
-
