+++ boilerplate/backend/python-drf/app/urls.py
@@ -1,8 +1,8 @@
-"""example URL Configuration
+"""app URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/3.2/topics/http/urls/
-Examples:
+apps:
 Function views
     1. Add an import:  from my_app import views
     2. Add a URL to urlpatterns:  path('', views.home, name='home')
@@ -16,7 +16,7 @@ Including another URLconf
 from django.contrib import admin
 from django.urls import path
 
-from example import views
+from app import views
 
 urlpatterns = [
     path('admin/', admin.site.urls),
